refactor(account-ms): add null check before deleting account or transaction by ID

+++ account-service/src/main/java/com/devsu/bank/account_service/service/AccountServiceImpl.java
@@ -55,6 +55,10 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public void deleteById(Long id) {
+        Boolean exists = accountRepository.existsById(id);
+        if (!exists) {
+            throw new AccountNotFoundException();
+        }
         accountRepository.deleteById(id);
     }
